azure: show login prompt correctly
Currently, Azure image does not show Scylla login prompt when logging
in to ssh.
This is because "azureuser" ($SSH_USERNAME) user is created at machine-image
building time, it's before we updated /etc/skel so it doesn't contain
scylla login prompt.
On Azure, username for SSH is not statically assigned, it is configurable
when launching instance.
So we don't need to keep "azureuser" ($SSH_USERNAME) user, we can drop
it when finishes machine-image creation, and newly created user will
have scylla login prompt.

Also, it's same on GCE, we don't need to keep $SSH_USERNAME user, we can
drop it too.